scripts/requirements-pythainlp.txt
fastapi
uvicorn[standard]
pythainlp
python-dotenv
